# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CommonPlace API — smart essentials platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products with optional search/filter
      parameters:
        - name: search
          in: query
          schema:
            type: string
            maxLength: 100
        - name: category
          in: query
          schema:
            type: string
            maxLength: 100
        - name: sponsored
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
            maximum: 100
        - name: limit
          in: query
          schema:
            type: integer
            default: 24
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}/similar:
    get:
      operationId: getSimilarProducts
      tags: [products]
      summary: Get similar/alternative products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Similar products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimilarProduct"

  # ── Orders ────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int32
            default: 1
            minimum: 1
            maximum: 100
        - name: limit
          in: query
          schema:
            type: integer
            format: int32
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created with checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCheckoutResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
            maximum: 100
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /admin/products:
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{id}:
    put:
      operationId: updateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    SimilarProduct:
      type: object
      properties:
        brand:
          type: string
        name:
          type: string
        price:
          type: number
        savings:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
      required: [brand, name, price]
    Product:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        wasPrice:
          type: number
          nullable: true
        bulkPrice:
          type: number
          nullable: true
        badge:
          type: string
          nullable: true
        badgeColor:
          type: string
          nullable: true
        savings:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        keywords:
          type: array
          items:
            type: string
          nullable: true
        similar:
          type: array
          items:
            $ref: "#/components/schemas/SimilarProduct"
          nullable: true
        isSponsored:
          type: boolean
        isFeatured:
          type: boolean
        inStock:
          type: boolean
      required: [id, category, name, price, isSponsored, isFeatured, inStock]
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [products, total, page, limit]
...
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
      required: [productId, name, quantity, price]
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        guestEmail:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, pending, paid, fulfilling, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        serviceFee:
          type: number
        total:
          type: number
        savings:
          type: number
          nullable: true
        shippingAddress:
          type: object
          nullable: true
          additionalProperties: true
        stripeSessionId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, status, items, subtotal, serviceFee, total, createdAt]
    CreateOrderBody:
      type: object
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: object
            properties:
              productId:
                type: integer
                format: int32
                minimum: 1
              quantity:
                type: integer
                format: int32
                minimum: 1
                maximum: 999
            required: [productId, quantity]
        guestEmail:
          type: string
          nullable: true
        autoRefill:
          type: boolean
        shippingAddress:
          type: object
          nullable: true
          properties:
            fullName:
              type: string
              maxLength: 100
            line1:
              type: string
              maxLength: 200
            line2:
              type: string
              maxLength: 200
              nullable: true
            city:
              type: string
              maxLength: 100
            state:
              type: string
              maxLength: 50
            zip:
              type: string
              maxLength: 20
            phone:
              type: string
              maxLength: 30
              nullable: true
          required: [fullName, line1, city, state, zip]
      required: [items]
    OrderCheckoutResponse:
      type: object
      properties:
        orderId:
          type: integer
        checkoutUrl:
          type: string
      required: [orderId, checkoutUrl]
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [orders, total, page, limit]
...
    CreateProductBody:
      type: object
      properties:
        category:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        wasPrice:
          type: number
          nullable: true
        bulkPrice:
          type: number
          nullable: true
        badge:
          type: string
          nullable: true
        badgeColor:
          type: string
          nullable: true
        savings:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        keywords:
          type: array
          items:
            type: string
          nullable: true
        isSponsored:
          type: boolean
        isFeatured:
          type: boolean
        inStock:
          type: boolean
      required: [category, name, price, isSponsored, isFeatured, inStock]